requirements.txt
Django==1.11.29
django-simple-history==2.12.0
djangorestframework==3.9.4
mysqlclient==1.4.6
Pillow==6.2.2
pytz==2022.2.1
six==1.16.0